feat: Extract cell configuration methods
ReminderViewController+CellConfiguration.swift
- Create new file as extension of ReminderViewController that have methods related with cell configuration.
